requirements.txt
google-genai
python-dotenv
groq
pandas
nltk
requests
openpyxl
beautifulsoup4
transformers
torch
tqdm
